Christmas time 2015 project
Kidz Poem Playlist, is an smartphone app project. The app makes a
dynamically growing list of items(poems) in the user interface when new
entry is added to the database i.e the app is connected to the database
and when new poem is added to the database along with the image,video
and name of Poem appears automatically within the app , without need of
updating the whole app ( .apk file).

The app is targeted to small kids and a trail and full version of the
app are developed, the source code of Trail version is available on my
Github account.